pyproject.toml
[project]
name = "llmops-travel-itinerary-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.34.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
